- extracted grid controller - tried to implement resizing behavior
