+++ day18fx/src/main/kotlin/aoc2015/day18/GridView.kt
@@ -1,36 +1,62 @@
 package aoc2015.day18
 
+import javafx.beans.binding.Bindings
 import javafx.beans.binding.When
 import javafx.beans.property.SimpleBooleanProperty
+import javafx.scene.control.Label
+import javafx.scene.layout.GridPane
 import javafx.scene.paint.Color
-import kotlinx.coroutines.*
 import tornadofx.*
-import kotlin.coroutines.CoroutineContext
 
 
 class GridView : View() {
 
     private val controller: GridController by inject()
 
-    override val root = vbox {
-        hbox {
-            button("play") {
-                action { controller.play() }
-            }
-            button("pause") {
-                action { controller.pause() }
+    private var grid: GridPane by singleAssign()
+
+    private var widthLabel: Label by singleAssign()
+    private var heightLabel: Label by singleAssign()
+
+    override val root = borderpane {
+        top {
+            hbox {
+                button("play") {
+                    action { controller.play() }
+                }
+                button("pause") {
+                    action { controller.pause() }
+                }
             }
         }
-        gridpane {
-            (0 until Grid.MAX_HEIGHT).forEach { y ->
-                row {
-                    (0 until Grid.MAX_WIDTH).forEach { x ->
-                        add(LightView(controller.gridProperties[Position(x, y)]
-                                ?: kotlin.error("no property found")))
+        center {
+            grid = gridpane {
+                background = Color.ALICEBLUE.asBackground()
+                (0 until Grid.MAX_HEIGHT).forEach { y ->
+                    row {
+                        (0 until Grid.MAX_WIDTH).forEach { x ->
+                            add(LightView(controller.lightAt(x, y)))
+                        }
                     }
                 }
             }
         }
+        bottom {
+            hbox {
+                widthLabel = label()
+                heightLabel = label()
+            }
+        }
+    }
+
+    init {
+        currentStage?.apply {
+            minWidth = 200.0
+            minHeight = 200.0
+        }
+        widthLabel.textProperty().bind(Bindings.selectString(currentStage?.widthProperty()))
+        heightLabel.textProperty().bind(Bindings.selectString(currentStage?.heightProperty()))
+
     }
 
     class LightView(
@@ -54,38 +80,3 @@ class GridView : View() {
     }
 }
 
-class GridController : Controller(), CoroutineScope {
-
-    override val coroutineContext: CoroutineContext
-        get() = Dispatchers.Default
-
-    private var grid = Day18.inputGrid
-    val gridProperties = Grid.positions().associateWith { SimpleBooleanProperty(grid[it]) }
-
-    private var job = Job().apply { cancel() }
-
-    fun play() {
-        job.cancel()
-        job = launch(Dispatchers.Default) {
-            while (isActive) {
-                delay(DELAY)
-                incrementGrid()
-            }
-        }
-    }
-
-    fun pause() {
-        job.cancel()
-    }
-
-    private fun incrementGrid() {
-        grid++
-        Grid.positions().forEach { pos ->
-            gridProperties[pos]?.set(grid[pos])
-        }
-    }
-
-    companion object {
-        private const val DELAY = 100L
-    }
-}
